aletheiadb::core::version

Trait EntityVersion

pub trait EntityVersion: TemporalVersion {
    // Required methods
    fn version_id(&self) -> VersionId;
    fn is_anchor(&self) -> bool;
    fn prev_version(&self) -> Option<VersionId>;
    fn set_prev_version(&mut self, version_id: Option<VersionId>);
    fn next_version(&self) -> Option<VersionId>;
    fn set_next_version(&mut self, version_id: Option<VersionId>);
    fn data_mut(&mut self) -> &mut VersionData;
}

Trait for version types that support common version chain operations.

This trait extends TemporalVersion with methods needed for version chain management, enabling generic implementations of version creation logic and reducing code duplication between node and edge version handling.

Trait Bounds

This trait requires TemporalVersion, which provides access to temporal intervals via temporal() and temporal_mut(). This allows generic code to both read and modify temporal properties.

Example

use crate::core::version::EntityVersion;

/// Count versions in a chain until reaching an anchor.
fn count_deltas_to_anchor<V: EntityVersion>(
    start_version: &V,
    get_version: impl Fn(VersionId) -> Option<&V>,
) -> usize {
    let mut count = 0;
    let mut current_id = start_version.prev_version();

    while let Some(vid) = current_id {
        if let Some(version) = get_version(vid) {
            if version.is_anchor() {
                break;
            }
            count += 1;
            current_id = version.prev_version();
        } else {
            break;
        }
    }
    count
}

Required Methods

fn version_id(&self) -> VersionId

Get the version’s unique identifier.

fn is_anchor(&self) -> bool

Check if this version is an anchor (full snapshot).

fn prev_version(&self) -> Option<VersionId>

Get the link to the previous version in the chain.

fn set_prev_version(&mut self, version_id: Option<VersionId>)

Set the link to the previous version in the chain.

fn next_version(&self) -> Option<VersionId>

Get the link to the next version in the chain.

fn set_next_version(&mut self, version_id: Option<VersionId>)

Set the link to the next version in the chain.

fn data_mut(&mut self) -> &mut VersionData

Get a mutable reference to the version data.