aletheiadb::core::version

Trait TemporalVersion

pub trait TemporalVersion {
    // Required methods
    fn temporal(&self) -> &BiTemporalInterval;
    fn temporal_mut(&mut self) -> &mut BiTemporalInterval;

    // Provided method
    fn close_transaction_time(&mut self, end_timestamp: Timestamp) -> Result<()> { ... }
}

Trait for version types that have a bi-temporal interval.

This trait provides a common interface for accessing and modifying the temporal interval of node and edge versions, reducing code duplication in operations that need to modify temporal properties.

Required Methods

fn temporal(&self) -> &BiTemporalInterval

Get a reference to the version’s bi-temporal interval.

fn temporal_mut(&mut self) -> &mut BiTemporalInterval

Get a mutable reference to the version’s bi-temporal interval.

Provided Methods

fn close_transaction_time(&mut self, end_timestamp: Timestamp) -> Result<()>

Close the transaction time of this version.

This marks the version as no longer being the “current knowledge” after the specified timestamp. Used when a version is superseded or deleted.