aletheiadb::embeddings

Function embed_file

pub async fn embed_file<T>(
    file_name: T,
    embedder: &Embedder,
    config: Option<&TextEmbedConfig>,
    adapter: Option<Box<dyn FnOnce(Vec<EmbedData>) + Send + Sync>>,
) -> Result<Option<Vec<EmbedData>>, Error>
where
    T: AsRef<Path>,

Embeds the text from a file using the specified embedding model.

Arguments

file_name - A string specifying the name of the file to embed.

embedder - A string specifying the embedding model to use. Valid options are “OpenAI”, “Jina”, “Clip”, and “Bert”.

config - An optional EmbedConfig object specifying the configuration for the embedding model.

‘adapter’ - An optional Adapter object to send the embeddings to a vector database.

Returns

A vector of EmbedData objects representing the embeddings of the file.

Errors

Returns a PyValueError if an invalid embedding model is specified.

Example

use embed_anything::{embed_file, embeddings::embed::EmbedderBuilder};
use embed_anything::config::TextEmbedConfig;

let embedder = EmbedderBuilder::new()
    .model_architecture("bert")
    .model_id(Some("sentence-transformers/all-MiniLM-L12-v2"))
    .from_pretrained_hf()?;
let config = TextEmbedConfig::default();
let embeddings = embed_file("document.pdf", &embedder, Some(&config), None).await?;