aletheiadb::index::vector::distributed

Trait VectorNodeClient

pub trait VectorNodeClient:
    Send
    + Sync
    + Debug {
    // Required methods
    fn node_id(&self) -> u16;
    fn is_healthy(&self) -> bool;
    fn add(&self, id: NodeId, vector: &[f32]) -> Result<()>;
    fn remove(&self, id: NodeId) -> Result<()>;
    fn search(&self, query: &[f32], k: usize) -> Result<Vec<(NodeId, f32)>>;
    fn len(&self) -> Result<usize>;
    fn health_check(&self) -> Result<()>;

    // Provided method
    fn is_empty(&self) -> Result<bool> { ... }
}

Trait for communicating with remote vector index nodes.

This trait abstracts the network communication, allowing for different implementations (HTTP, gRPC, in-process mock for testing).

Required Methods

fn node_id(&self) -> u16

Get the node ID this client connects to.

fn is_healthy(&self) -> bool

Check if the connection is healthy.

fn add(&self, id: NodeId, vector: &[f32]) -> Result<()>

Add a vector to the remote index.

fn remove(&self, id: NodeId) -> Result<()>

Remove a vector from the remote index.

fn search(&self, query: &[f32], k: usize) -> Result<Vec<(NodeId, f32)>>

Search for k-nearest neighbors on the remote index.

fn len(&self) -> Result<usize>

Get the number of vectors on the remote index.

fn health_check(&self) -> Result<()>

Perform a health check.

Provided Methods

fn is_empty(&self) -> Result<bool>

Check if the remote index is empty.