aletheiadb

Module prelude

Prelude for convenient imports of common types and traits. AletheiaDB Prelude

The prelude module re-exports the most commonly used traits and types to make starting with AletheiaDB easier.

Usage

use aletheiadb::prelude::*;

Re-exports

pub use crate::AletheiaDB;

pub use crate::api::ReadOps;

pub use crate::api::ReadTransaction;

pub use crate::api::WriteOps;

pub use crate::api::WriteTransaction;

pub use crate::core::error::Error;

pub use crate::core::error::Result;

pub use crate::core::id::EdgeId;

pub use crate::core::id::NodeId;

pub use crate::core::id::VersionId;

pub use crate::core::interning::InternedString;

pub use crate::core::property::PropertyMap;

pub use crate::core::property::PropertyMapBuilder;

pub use crate::core::property::PropertyValue;

pub use crate::core::temporal::BiTemporalInterval;

pub use crate::core::temporal::TimeRange;

pub use crate::core::temporal::Timestamp;

pub use crate::core::temporal::time;

pub use crate::storage::wal::DurabilityMode;

pub use crate::storage::wal::WriteOptions;

Macros

properties

Macro for creating property maps with a convenient syntax.