aletheiadb::semantic_search

Module concept_algebra

Concept Algebra: Vector Arithmetic for Nodes.

This module allows performing semantic arithmetic on nodes. It treats nodes as their underlying vector embeddings and allows operations like:

Addition: Concept(A) + Concept(B) (Composition)

Subtraction: Concept(A) - Concept(B) (Removal of meaning)

Analogy: Concept(A) - Concept(B) + Concept(C) (e.g. “King” - “Man” + “Woman” = “Queen”)

Mean: Centroid of multiple concepts.

Quick Start

use aletheiadb::AletheiaDB;
use aletheiadb::semantic_search::concept_algebra::ConceptAlgebra;
use aletheiadb::core::id::NodeId;

let db = AletheiaDB::new()?;
// ... (assume db is populated with nodes and vectors) ...
let king = NodeId::new(1).unwrap();
let man = NodeId::new(2).unwrap();
let woman = NodeId::new(3).unwrap();

let algebra = ConceptAlgebra::new(&db);

// Perform analogy: King - Man + Woman = ?
let results = algebra.analogy(king, man, woman, 5)?;

for (node_id, score) in results {
    println!("Found node {} with score {}", node_id, score);
}

Structs

ConceptAlgebra

A tool for performing vector arithmetic on graph nodes.